Extent artifact argument & add "additional-setup" input & some fixes (#72)
* Extent artifact argument & add "additional-setup" input

* Solve garrysmod_override never being synced & fix tar complaining

* Fail workflow on GLuaTest failure

* Readd status & exit if it didn't succeed

* Enter the right directory when pulling the docker compose yml

* Also fail if the gluatest_clean_exit.txt doesn't even exists

* Try to solve self-tests & let them run on every push too

* Move this out of docker as the last pr forgot this

* Remove old convar as it doesn't exist anymore

* Add missing elif's to entrypoint

* Try to get PR runs to work again

* Rename job to run_tests

* Limit the push trigger to the main branch


* Fix ::error:: reporting not working

* Don't include invalid source files

* Bring back colors and solve a bug

---------

+++ docker/entrypoint.sh
@@ -15,15 +15,17 @@ rm -rf _tmp_ssh _tmp_https
 
 # Copy the overrides overtop the server files
 echo "Copying serverfiles overrides..."
-rsync --verbose --archive $home/serverfiles_override/ $gmodroot/
+rsync --verbose --archive $home/garrysmod_override/ $server/
 
 # Any additional files
-cp $home/_gluatest_artifacts/_gluatest_artifacts/* $gmodroot/
+if [ -n "$(ls -A $home/_gluatest_artifacts/_gluatest_artifacts/)" ]; then # Only execute if there are any artifacts or else tar will complain
+    cp $home/_gluatest_artifacts/_gluatest_artifacts/* $gmodroot/
 
-for file in $gmodroot/*.tar.gz; do \
-    tar --extract --verbose --ungzip --file="$file" --directory="$gmodroot" && \
-    rm --force --verbose "$file"; \
-done
+    for file in $gmodroot/*.tar.gz; do \
+        tar --extract --verbose --ungzip --file="$file" --directory="$gmodroot" && \
+        rm --force --verbose "$file"; \
+    done
+fi
 
 if [ -f "$gmodroot/custom_requirements.txt" ]; then
     echo "Appending custom requirements"
@@ -138,9 +140,22 @@ echo "GMOD_BRANCH: $GMOD_BRANCH"
 if [ "$GMOD_BRANCH" = "x86-64" ]; then
     echo "Starting 64-bit server"
     unbuffer timeout "$timeout" "$gmodroot"/srcds_run_x64 "$srcds_args"
+elif [ "$GMOD_BRANCH" = "prerelease" ]; then
+    echo "Starting 32-bit prerelease server"
+    unbuffer timeout "$timeout" "$gmodroot"/srcds_run "$srcds_args"
+elif [ "$GMOD_BRANCH" = "dev" ]; then
+    echo "Starting 32-bit dev server"
+    unbuffer timeout "$timeout" "$gmodroot"/srcds_run "$srcds_args"
 else
     echo "Starting 32-bit server"
     unbuffer timeout "$timeout" "$gmodroot"/srcds_run "$srcds_args"
+fi
+
+status=$?
+
+if [ "$status" -ne 0 ]; then
+    echo "::error:: Something went wrong! - Failing workflow"
+    exit "$status"
 fi
 
 if [ -f "$gmodroot/debug.log" ]; then
@@ -150,11 +165,9 @@ if [ -f "$gmodroot/debug.log" ]; then
 	exit 1
 fi
 
-status=$?
-
-if [ "$(cat $server/data/gluatest_clean_exit.txt)" = "false" ]; then
+if [ ! -f "$server/data/gluatest_clean_exit.txt" ] || [ "$(cat $server/data/gluatest_clean_exit.txt)" = "false" ]; then
     echo "::warning:: Test runner did not exit cleanly. Test results unavailable!"
-    exit "$status"
+    exit 1 # This should never happen normally, GLuaTest probably had an error while executing so we should fail.
 fi
 
 if [ -s "$server/data/gluatest_failures.json" ]; then
